rename: base_cache -> base_prefix_cache
